added dependency check with webbler forms

+++ commonlib/pages/attributes.php
@@ -1,6 +1,13 @@
 <?
+$types = array('textline','checkbox','checkboxgroup','radio','select',"hidden","textarea");
+$formtable_exists = Sql_Table_exists("formfield");
 
 ob_end_flush();
+#foreach ($_POST as $key => $val) {
+#  print "$key = ".print_r($val)."<br/>";
+#}
+#return;
+
 print '<script language="Javascript" src="js/progressbar.js" type="text/javascript"></script>';
 if (isset($_POST["action"]) && $_POST["action"] == "Save Changes") {
   if (isset($_POST["name"])) {
@@ -8,54 +15,55 @@ if (isset($_POST["action"]) && $_POST["action"] == "Save Changes") {
     while (list($id,$val) = each ($_POST["name"])) {
       if (!$id && isset($_POST["name"][0]) && $_POST["name"][0] != "") {
         # it is a new one
-        $lc_name = substr(preg_replace("/\W/","", strtolower($_POST["name"][0])),0,10);
-        if ($lc_name == "") { Fatal_Error("Name cannot be empty: $lc_name");return; }
-        Sql_Query("select * from {$tables['attribute']} where tablename = \"$lc_name\"");
-        if (Sql_Affected_Rows()) { Fatal_Error("Name is not unique enough"); return; }
+        $lc_name = getNewAttributeTablename($_POST["name"][0]);        
         if ($lc_name == "email") { Fatal_Error("Email is a system attribute, you cannot add it manually"); return; }
 
+        #print "New attribute: ".$_POST["name"][0]."<br/>";
         $query = sprintf('insert into %s (name,type,listorder,default_value,required,tablename) values("%s","%s",%d,"%s",%d,"%s")',
         $tables["attribute"],addslashes($_POST["name"][0]),$_POST["type"][0],$_POST["listorder"][0],addslashes($_POST["default"][0]),$_POST["required"][0],$lc_name);
         Sql_Query($query);
         $insertid = Sql_Insert_id();
 
         # text boxes and hidden fields do not have their own table
-        if ($type[$id] != "textline" && $type["id"] != "hidden") {
+        if ($_POST["type"][$id] != "textline" && $_POST["type"]["id"] != "hidden") {
           $query = "create table $table_prefix"."listattr_$lc_name (id integer not null primary key auto_increment, name varchar(255) unique,listorder integer default 0)";
           Sql_Query($query);
         } else {
           # and they cannot currently be required, changed 29/08/01, insert javascript to require them, except for hidden ones :-)
-          if ($type["id"] == "hidden")
+          if ($_POST["type"]["id"] == "hidden")
             Sql_Query("update {$tables['attribute']} set required = 0 where id = $insertid");
         }
-        if ($type[$id] == "checkbox") {
+        if ($_POST["type"][$id] == "checkbox") {
           # with a checkbox we know the values
 #          Sql_Query('insert into '.$table_prefix.'listattr_'.$lc_name.' (name) values("Checked")');
 #          Sql_Query('insert into '.$table_prefix.'listattr_'.$lc_name.' (name) values("Unchecked")');
           # we cannot "require" checkboxes, that does not make sense
           Sql_Query("update {$tables['attribute']} set required = 0 where id = $insertid");
         }
-        if ($type[$id] == "checkboxgroup")
+        if ($_POST["type"][$id] == "checkboxgroup")
           Sql_Query("update {$tables['attribute']} set required = 0 where id = $insertid");
 
         # fix all existing users to have a record for this attribute, even with empty data
         $req = Sql_Query("select id from {$tables["user"]}");
         while ($row = Sql_Fetch_Row($req)) {
-          Sql_Query(sprintf('insert ignore into %s (attributeid,userid) values(0,%d)',
-            $tables["user_attribute"],$row[0]));
+          Sql_Query(sprintf('insert ignore into %s (attributeid,userid) values(%d,%d)',
+            $tables["user_attribute"],$insertid,$row[0]));
         }
       } elseif ($_POST["name"][$id] != "") {
         # it is a change
         # get the original type
 
         $req = Sql_Fetch_Row_Query("select type,tablename from {$tables['attribute']} where id = $id");
-        switch($req[0]) {
+        $existingtype = $req[0];
+        #print "Existing attribute: ".$_POST["name"][$id]." new type:".$_POST["type"][$id]." existing type: ".$req[0]."<br/>";
+        
+        if ($_POST["type"][$id] != $existingtype)
+        switch ($existingtype) {
           case "textline":case "hidden":
-            if ($type[$id] == "hidden" || $type[$id] == "textline") break;
+            if ($_POST["type"][$id] == "hidden" || $_POST["type"][$id] == "textline") break;
             # we are turning a hidden or textline field into a radio,checkbox,checkboxgroup or select
-            if ($type[$id] != "checkbox") {
-              $lc_name = substr(preg_replace("/\W/","", strtolower($req[1])),0,10);
-              if ($lc_name == "") Fatal_Error("Name cannot be empty: $lc_name");
+            if ($_POST["type"][$id] != "checkbox") {
+              $lc_name = getNewAttributeTablename($req[1]);
               Sql_Query("create table $table_prefix"."listattr_$lc_name (id integer not null primary key auto_increment, name varchar(255) unique,listorder integer default 0)");
               $attreq = Sql_Query("select distinct value from {$tables['user_attribute']} where attributeid = $id");
               while ($row = Sql_Fetch_Row($attreq)) {
@@ -76,7 +84,7 @@ if (isset($_POST["action"]) && $_POST["action"] == "Save Changes") {
             }
             break;
           case "radio":case "select": case "checkbox":
-            if ($type[$id] != "hidden" && $type[$id] != "textline") break;
+            if ($_POST["type"][$id] != "hidden" && $_POST["type"][$id] != "textline") break;
             # we are turning a radio,select or checkbox into a hidden or textline field
             $valuereq = Sql_Query("select id,name from $table_prefix"."listattr_$req[1]");
             while ($row = Sql_Fetch_Row($valuereq))
@@ -84,14 +92,14 @@ if (isset($_POST["action"]) && $_POST["action"] == "Save Changes") {
             Sql_Query("drop table $table_prefix"."listattr_$req[1]");
             break;
           case "checkboxgroup":
-            if ($type[$id] == "hidden" || $type[$id] == "textline") {
+            if ($_POST["type"][$id] == "hidden" || $_POST["type"][$id] == "textline") {
             	# we are changing a checkbox group into a hidden or textline
               # take the first value!
               $valuereq = Sql_Query("select id,name from $table_prefix"."listattr_$req[1]");
               while ($row = Sql_Fetch_Row($valuereq))
                 Sql_Query("update {$tables['user_attribute']} set value = \"$row[1]\" where attributeid = $id and value like \"$row[0]%\"");
 	            Sql_Query("drop table if exists $table_prefix"."listattr_$req[1]");
-            } elseif ($type[$id] == "radio" || $type[$id] == "select") {
+            } elseif ($_POST["type"][$id] == "radio" || $_POST["type"][$id] == "select") {
               $valuereq = Sql_Query("select userid,value from {$tables["user_attribute"]} where attributeid = $id");
               # take the first value!
               while ($row = Sql_Fetch_Row($valuereq)) {
@@ -102,7 +110,7 @@ if (isset($_POST["action"]) && $_POST["action"] == "Save Changes") {
             break;
         }
         $query = sprintf('update %s set name = "%s" ,type = "%s" ,listorder = %d,default_value = "%s" ,required = %d where id = %d',
-        $tables["attribute"],addslashes($_POST["name"][$id]),$type[$id],$listorder[$id],$default[$id],$required[$id],$id);
+        $tables["attribute"],addslashes($_POST["name"][$id]),$_POST["type"][$id],$listorder[$id],$default[$id],$required[$id],$id);
         Sql_Query($query);
       }
     }
@@ -110,17 +118,31 @@ if (isset($_POST["action"]) && $_POST["action"] == "Save Changes") {
   }
 } elseif (isset($_POST["tagaction"]) && is_array($_POST["tag"])) {
 	ksort($_POST["tag"]);
-	if ($_POST["tagaction"] == "Delete") {
+  if ($_POST["tagaction"] == "Delete") {
     while (list($k,$id) = each ($_POST["tag"])) {
-    	print "Deleting ".$id."<br/>";
-      $row = Sql_Fetch_Row_Query("select tablename,type from {$tables['attribute']} where id = $id");
-      Sql_Query("drop table if exists $table_prefix"."listattr_$row[0]");
-      Sql_Query("delete from {$tables['attribute']} where id = $id");
-      # delete all user attributes as well
-      Sql_Query("delete from {$tables['user_attribute']} where attributeid = $id");
+      # check for dependencies
+      if ($formtable_exists) {
+        $req = Sql_Query("select * from formfield where attribute = $id");
+        $candelete = !Sql_Affected_Rows();
+      } else {
+        $candelete = 1;
+      }
+      if ($candelete) {
+        print "Deleting ".$id."<br/>";
+        $row = Sql_Fetch_Row_Query("select tablename,type from {$tables['attribute']} where id = $id");
+        Sql_Query("drop table if exists $table_prefix"."listattr_$row[0]");
+        Sql_Query("delete from {$tables['attribute']} where id = $id");
+        # delete all user attributes as well
+        Sql_Query("delete from {$tables['user_attribute']} where attributeid = $id");
+      } else {
+        print Error("Cannot delete attribute $id, it is being used by the following forms:<br/>");
+        while ($row = Sql_Fetch_Array($req)) {
+          print PageLink2("editelements&id=".$row["form"]."&option=edit_elements&pi=formbuilder","form ".$row["form"]."")."<br/>\n";
+        }
+      }
     }
  	} elseif ($_POST["tagaction"] == "Merge") {
-  	$first = array_shift($_POST["tag"]);
+    $first = array_shift($_POST["tag"]);
     $firstdata = Sql_Fetch_Array_Query(sprintf('select * from %s where id = %d',$tables["attribute"],$first));
     if (!sizeof($_POST["tag"])) {
     	print Error("cannot merge just one attribute");
@@ -128,10 +150,24 @@ if (isset($_POST["action"]) && $_POST["action"] == "Save Changes") {
     	$cbg_initiated = 0;
     	foreach ($_POST["tag"] as $attid) {
       	print "Merging $attid into $first<br/>";
+        
 		    $attdata = Sql_Fetch_Array_Query(sprintf('select * from %s where id = %d',$tables["attribute"],$attid));
         if ($attdata["type"] != $firstdata["type"]) {
         	print Error("Can only merge attributes of the same type");
         } else {
+          # debugging: check values for every user. This is very memory demanding, so you'll need to 
+          # add loads of memory to actually use it.
+          $before = array();
+          $second = array();
+          $after = array();
+          $req = Sql_Query(sprintf('select * from %s where attributeid = %d',$tables["user_attribute"],$first));
+          while ($row = Sql_Fetch_Array($req)) {
+            $before[$row["userid"]] = $row["value"];
+          }
+          $req = Sql_Query(sprintf('select * from %s where attributeid = %d',$tables["user_attribute"],$attid));
+          while ($row = Sql_Fetch_Array($req)) {
+            $second[$row["userid"]] = $row["value"];
+          }
           $valuestable = sprintf('%slistattr_%s',$table_prefix,$firstdata["tablename"]);
         	if ($firstdata["type"] == "checkbox" && !$cbg_initiated) {
             # checkboxes are merged into a checkbox group
@@ -155,8 +191,12 @@ if (isset($_POST["action"]) && $_POST["action"] == "Save Changes") {
             	# we can just keep the data and mark it as the first attribute
               Sql_query(sprintf('update ignore %s set attributeid = %d where attributeid = %d',
                 $tables["user_attribute"],$first,$attid),1);
+              # delete the ones that didn't copy across, because there was a value already  
               Sql_query(sprintf('delete from %s where id = %d',
                 $tables["attribute"],$attid));
+              # mark forms to use the merged attribute
+              if ($formtable_exists)
+                Sql_Query(sprintf('update formfield set attribute = %d where attribute = %d',$first,$attid),1);
               break;
             case "radio":
             case "select":
@@ -183,6 +223,9 @@ if (isset($_POST["action"]) && $_POST["action"] == "Save Changes") {
               Sql_Query(sprintf('drop table %s',$table_prefix."listattr_".$attdata["tablename"]),1);
               Sql_Query(sprintf('delete from %s where id = %d',
                 $tables["attribute"],$attid));
+              # mark forms to use the merged attribute
+              if ($formtable_exists)
+                Sql_Query(sprintf('update formfield set attribute = %d where attribute = %d',$first,$attid),1);
               break;
             case "checkbox":
               $exists = Sql_Fetch_row_Query(sprintf('select id from %s where name = "%s"',
@@ -198,12 +241,23 @@ if (isset($_POST["action"]) && $_POST["action"] == "Save Changes") {
                 $tables["user_attribute"],$val_index,$first));
               Sql_Query(sprintf('delete from %s where id = %d',
                 $tables["attribute"],$attid));
+              # mark forms to use the merged attribute
+              if ($formtable_exists)
+                Sql_Query(sprintf('update formfield set attribute = %d where attribute = %d',$first,$attid),1);
               break;
             case "checkboxgroup":
             	# hmm, this is a tricky one.
              	print Error("Sorry, merging of checkbox groups is not implemented yet");
               break;
          	}
+          
+          $req = Sql_Query(sprintf('select * from %s where attributeid = %d',$tables["user_attribute"],$first));
+          while ($row = Sql_Fetch_Array($req)) {
+            $after[$row["userid"]] = $row["value"];
+          }
+          foreach ($before as $userid => $value) {
+            printf("\n".'<br/>%d before -> %s and %s<br/>after ->%s',$userid,$value,$second[$userid],$after[$userid]);
+          }
         }
       }
     }
@@ -235,15 +289,21 @@ while ($row = Sql_Fetch_array($res)) {
 	$c++;
   ?>
   <table border=1>
-  <tr><td colspan=2>Attribute:<? echo $row["id"] ?></td><td colspan=2>Tag <input type="checkbox" name="tag[<?=$c?>]" value="<?=$row["id"]?>"></td></tr>
+  <tr><td colspan=2>Attribute:<? echo $row["id"];
+if ($formtable_exists) {
+  sql_query("select * from formfield where attribute = ".$row["id"]);
+  print "  (used in ".Sql_affected_rows()." forms)";
+}
+    
+?>  </td><td colspan=2>Tag <input type="checkbox" name="tag[<?=$c?>]" value="<?=$row["id"]?>"></td></tr>
+    
   <tr><td colspan=2>Name: </td><td colspan=2><input type=text name="name[<? echo $row["id"]?>]" value="<? echo htmlspecialchars(stripslashes($row["name"])) ?>" size=40></td></tr>
   <tr><td colspan=2>Type: </td><td colspan=2><!--input type=hidden name="type[<?=$row["id"]?>]" value="<?=$row["type"]?>"><?=$row["type"]?>-->
 
   <select name="type[<?=$row["id"]?>]" onChange="warn();">
 <?
-	 $types = array('textline','checkbox','checkboxgroup','radio','select',"hidden","textarea");
-   while (list($key,$val) = each($types)) {
-     printf('<option value="%s" %s>%s',$val,$val == $row["type"] ? "selected": "",$val);
+   foreach($types as $key => $val) {
+     printf('<option value="%s" %s>%s</option>',$val,$val == $row["type"] ? "selected": "",$val);
    }
 ?>
    </select>
@@ -271,9 +331,8 @@ while ($row = Sql_Fetch_array($res)) {
 <tr><td colspan=2>Name: </td><td colspan=2><input type=text name="name[0]" value="" size=40></td></tr>
 <tr><td colspan=2>Type: </td><td colspan=2><select name="type[0]">
 <?
-$types = array('textline','checkbox','checkboxgroup','radio','select',"hidden","textarea");
-while (list($key,$val) = each($types)) {
-  printf('<option value="%s" %s>%s',$val,"",$val);
+foreach($types as $key => $val) {
+  printf('<option value="%s" %s>%s</option>',$val,"",$val);
 }
 ?>
 </select></td></tr>
